+++ src/uel/errors/uelbaseexception.py
@@ -1,9 +1,10 @@
 __all__ = ["UELBaseException"]
 
 
-class UELBaseException:
+class UELBaseException(Exception):
     def __init__(self, error_message: str):
         self.error_message = error_message
 
     def __str__(self) -> str:
-        return f"{self.__class__.__name__}:{self.error_message}"
+        msg = self.error_message
+        return f"{self.__class__.__module__}.{self.__class__.__name__}: {msg}"