+++ src/uel/hook/excepthook.py
@@ -1,0 +1,12 @@
+import sys
+from uel.errors.uelbaseexception import UELBaseException
+from uel.errors.ueerror_setstring import UEErr_SetString
+
+default_excepthook = sys.excepthook
+
+def excepthook(exctype, value, traceback):
+    if issubclass(exctype, UELBaseException):
+        UEErr_SetString(value)
+        return
+    
+    default_excepthook(exctype, value, traceback)