+++ src/uel/hook/initiation.py
@@ -1,0 +1,6 @@
+import sys
+
+from uel.hook.excepthook import excepthook
+
+def initialization():
+    sys.excepthook = excepthook